backend/requirements.txt
fastapi[standard]
langchain
langchain_chroma
langchain_core
langchain_openai
python-dotenv
uvicorn